revise test_mac.yml

+++ test_mac/test_pow.m
@@ -3,4 +3,4 @@ function y = test_pow(x)
 mex('-setup', '-v', 'Fortran');
 mex('-v', 'pow.F');
 y = pow(x);
-assert(y == x^x, 'pow failed')
+assert(y == x^int64(x), 'pow failed')
